♻️ refactor: 채팅 도메인 리팩토링
