+++ src/pages/coffeechat/groupchatpage/index.tsx
@@ -1,0 +1,5 @@
+import GroupChatContainer from '@/pages/coffeechat/groupchatpage/GroupChatContainer';
+
+export default function GroupChatPage() {
+  return <GroupChatContainer />;
+}
