+++ src/routes/main/CoffeeChatRoutes.tsx
@@ -1,7 +1,7 @@
 import { Routes, Route } from 'react-router-dom';
 import CoffeeChatPage from '@/pages/coffeechat/coffeechatpage';
 import CoffeeChatDetailPage from '@/pages/coffeechat/coffeechatdetailpage';
-import GroupChatPage from '@/pages/coffeechat/GroupChatPage';
+import GroupChatPage from '@/pages/coffeechat/groupchatpage';
 import CoffeeChatReviewPage from '@/pages/coffeechat/CoffeeChatReviewPage';
 import NotFoundPage from '@/pages/common/NotFoundPage';
